backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.29
alembic>=1.13.1
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
emails>=0.6
jinja2>=3.1.3
email-validator>=2.1.0
psycopg2-binary>=2.9.9
argon2-cffi>=23.1.0
google-generativeai>=0.5.0
openai>=1.20.0
python-dotenv>=1.0.0
google-auth>=2.29.0
requests>=2.31.0
firebase-admin>=6.5.0
stripe>=8.0.0
mercadopago>=2.2.0
